avoid unitialized warnings when resolving extra compiler/link flags in xs_ok
xs_ok merges the compile and link flags passed to it with those
provided by ExtUtils::CppGuess. It assumes that if there are extra
flags returned by ExtUtils::CppGuess, then there are also extra flags
passed to xs_ok.

On some systems, this assumption does not hold, and results in
undefined values being passed to Text::ParseWords::shellwords when parsing
extra flags passed to xs_ok, causing warnings such as

  use of uninitialized value $line in substitution (s///) at .../Text/ParseWords.pm line 21.

to be emitted.